# lib/api-zod/src/generated/types/adminClient.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ditheto Accountants administration API
 * OpenAPI spec version: 0.2.0
 */
import type { AdminClientBranch } from './adminClientBranch';
import type { AdminClientOdooSyncStatus } from './adminClientOdooSyncStatus';

export interface AdminClient {
  id: number;
  name: string;
  email: string;
  phone: string;
  company: string;
  branch: AdminClientBranch;
  odooSyncStatus: AdminClientOdooSyncStatus;
  createdAt: Date;
}