# lib/api-zod/src/generated/types/clientDocument.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ditheto Accountants administration API
 * OpenAPI spec version: 0.2.0
 */
import type { DocumentCategory } from './documentCategory';
import type { DocumentStatus } from './documentStatus';

export interface ClientDocument {
  id: number;
  clientId: number;
  name: string;
  category: DocumentCategory;
  status: DocumentStatus;
  objectPath: string;
  mimeType: string;
  size: number;
  /** @nullable */
  expiresAt?: Date | null;
  uploadedBy: string;
  createdAt: Date;
  updatedAt: Date;
}