# lib/api-zod/src/generated/types/clientDocumentUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ditheto Accountants administration API
 * OpenAPI spec version: 0.2.0
 */
import type { DocumentCategory } from './documentCategory';
import type { DocumentStatus } from './documentStatus';

export interface ClientDocumentUpdate {
  /** @minLength 1 */
  name?: string;
  category?: DocumentCategory;
  status?: DocumentStatus;
  objectPath?: string;
  mimeType?: string;
  /** @minimum 1 */
  size?: number;
  /** @nullable */
  expiresAt?: Date | null;
}