# lib/api-zod/src/generated/types/documentCategory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ditheto Accountants administration API
 * OpenAPI spec version: 0.2.0
 */

export type DocumentCategory = typeof DocumentCategory[keyof typeof DocumentCategory];


export const DocumentCategory = {
  ID: 'ID',
  Tax_Clearance: 'Tax Clearance',
  IRP5: 'IRP5',
  Bank_Statements: 'Bank Statements',
  Contracts: 'Contracts',
  SARS_Correspondence: 'SARS Correspondence',
  Other: 'Other',
} as const;